Split Filter and FilterSet classes
